Implemented very basic build card creation.
It literally just rerenders the app as a PNG and downloads to the client's system. I'll be looking into how to make build cards that are more suitable for their intended purpose.

It's also currently broken due to a weird html2canvas library bug that causes white offsets to appear. It seems to always happen when the scrollbar appears, but it sometimes also appears without the scrollbar. Really weird.